Keep GitHub config tests lint-clean
